requirements.txt
flask
line-bot-sdk
google-generativeai>=0.7.0
gunicorn